% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{billionaires}
\alias{billionaires}
\title{Billionaires data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 500 rows and 6 columns.
}
\source{
\url{https://www.bloomberg.com/billionaires/}
}
\usage{
billionaires
}
\description{
Contains 500 billionaires' name, country, industry, and total net worth. The data is collected in Feb. 8, 2018.
}
\keyword{datasets}
