% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglorenz-package.r, R/stat_lorenz.R,
%   R/stat_lorenz_generalized.R
\docType{package}
\name{gglorenz}
\alias{gglorenz}
\alias{StatLorenz}
\alias{StatLorenzGeneralized}
\title{gglorenz: Plotting Lorenz Curve with ggplot2}
\description{
The package provides statistical transformations for plotting
empirical ordinary Lorenz curve and generalized Lorenz curve.
}
\author{
JJ Chen
}
\keyword{datasets}
